moved ConfigProvider methods to new abstract class as suggested
